+++ persistence/src/main/java/cz/muni/fi/pa165/sportsClub/dao/ManagerDaoImpl.java
@@ -1,6 +1,7 @@
 package cz.muni.fi.pa165.sportsClub.dao;
 
 import javax.persistence.EntityManager;
+import javax.persistence.NoResultException;
 import javax.persistence.PersistenceContext;
 import javax.persistence.TypedQuery;
 
@@ -21,28 +22,41 @@ public class ManagerDaoImpl implements ManagerDao{
 
 	@Override
 	public void createManager(Manager m) {
+		if(m == null){
+			throw new IllegalArgumentException("Argument can not be null");
+		}
 		em.persist(m);
 	}
 
 	@Override
 	public Manager getManagerById(Long id) {
+
 		return em.find(Manager.class, id);
 	}
 
 	@Override
 	public Manager getManagerByEmail(String email) {
-		TypedQuery<Manager> query = em.createQuery("select m from Manager m where email like :email", Manager.class);
-		query.setParameter("email", email);
-		return query.getSingleResult();
+		try{
+			return em.createQuery("select m from Manager m where email like :email", Manager.class).
+			setParameter("email", email).getSingleResult();
+		} catch (NoResultException nrf) {
+			return null;
+		}
 	}
 
 	@Override
 	public void updateManager(Manager m) {
+		if(m == null){
+			throw new IllegalArgumentException("Argument can not be null");
+		}
 		em.merge(m);
 	}
 
 	@Override
-	public void deleteManager(Manager m) {
+	public void deleteManager(Manager m){
+		if(m == null){
+			throw new IllegalArgumentException("Argument can not be null");
+		}
 		em.remove(em.merge(m));
 	}
 }
